Provide a warning of potentially unknown Slurm params
In some recent Slurm versions, the Slurm runtime is inserting
custom arguments to the PRRTE launcher's `srun` cmd line without
the user being aware of it. In many cases, this may not be a
problem - but in some cases (where the user or the system
admin needs/wants particular cmd line arguments used) this can
cause problems as it happens silently, without the user being
aware of it.

Make this visible when it happens, and provide a mechanism by
which the user/admin can override it. Provide a fairly long
help message explaining what happened and offering advice on
resolution, along with a param for disabling the warning. Add
a param for overriding the "args" param if necessary, along
with a caution as to possible consequences.

Signed-off-by: Ralph Castain <rhc@pmix.org>